refactor: translate new message, fix bugs

+++ app/Http/Controllers/API/AuthController.php
@@ -93,21 +93,25 @@ class AuthController extends Controller
 
     public function sendPasswordResetRequest(PasswordResetEmailRequest $request): JsonResponse
     {
-        $token = Str::random(40);
-        $data['token'] = $token;
-        $data['email'] = $request->email;
+        if(User::where('email', $request->email)->firstOrFail()->google_id) {
+            $token = Str::random(40);
+            $data['token'] = $token;
+            $data['email'] = $request->email;
 
-        Mail::send('email.password-reset', ['data' => $data], function ($message) use ($data) {
-            $message->to($data['email'])->subject('Reset Password');
-        });
+            Mail::send('email.password-reset', ['data' => $data], function ($message) use ($data) {
+                $message->to($data['email'])->subject('Reset Password');
+            });
 
-        if(ChangePassword::where('email', $request->email)->first()) {
-            ChangePassword::where('email', $request->email)->first()->delete();
+            if(ChangePassword::where('email', $request->email)->first()) {
+                ChangePassword::where('email', $request->email)->first()->delete();
+            }
+
+            ChangePassword::create(['email' => $request->email, 'token' => $token]);
+
+            return response()->json(['message' => __('messages.email_confirmation_sent_for_reset_password')]);
         }
 
-        ChangePassword::create(['email' => $request->email, 'token' => $token]);
-
-        return response()->json(['message' => __('messages.email_confirmation_sent_for_reset_password')]);
+        return response()->json(['message' => __('messages.invalid_credentials')]);
     }
 
 
@@ -116,7 +120,7 @@ class AuthController extends Controller
         $changePasswordModel = ChangePassword::where('token', $request->token)->first();
         if($changePasswordModel) {
             if($changePasswordModel->expires_at > Carbon::now()) {
-                return redirect()->away(env('FRONTEND_URL').`/reset-password/$request->token`);
+                return redirect()->away(env('FRONTEND_URL')."/reset-password/$request->token");
             }
             $changePasswordModel->delete();
             return redirect()->away(env('FRONTEND_URL').'/expired');
